+++ app/domains/prozesskostenhilfe/services/pdf/D_angehoerige.ts
@@ -1,6 +1,6 @@
-import type { PkhPdfFillFunction } from "~/domains/prozesskostenhilfe/pdf";
 import { getFillUnterhalt } from "~/domains/shared/pdf/unterhaltHelpers";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "../../../services/pdf/attachment";
+import { PkhPdfFillFunction } from ".";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 export const ATTACHMENT_DESCRIPTION_SECTION_D =
   "FELD D: Angehörige, denen Sie Bar- oder Naturalunterhalt gewähren";
