+++ app/domains/prozesskostenhilfe/services/pdf/__test__/A_person.test.ts
@@ -3,6 +3,7 @@ import {
   type ProzesskostenhilfePDF,
 } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
+import { maritalDescriptionMapping } from "~/domains/shared/pdf/maritalDescriptionMapping";
 import {
   ANSCHRIFT_FIELD_MAX_CHARS,
   BERUF_FIELD_MAX_CHARS,
@@ -12,9 +13,8 @@ import {
   fillPerson,
   GESETZLICHERVERTRETER_FIELD_MAX_CHARS,
   NAME_VORNAME_FIELD_MAX_CHARS,
-} from "~/domains/prozesskostenhilfe/pdf/A_person";
-import { maritalDescriptionMapping } from "~/domains/shared/pdf/maritalDescriptionMapping";
-import { SEE_IN_ATTACHMENT_DESCRIPTION } from "../../../../services/pdf/attachment";
+} from "../A_person";
+import { SEE_IN_ATTACHMENT_DESCRIPTION } from "~/services/pdf/attachment";
 
 let pdfParams: ProzesskostenhilfePDF;
 const userData: ProzesskostenhilfeFormularContext = {
