+++ app/domains/prozesskostenhilfe/services/pdf/__test__/C_unterhaltspflichtige_person.test.ts
@@ -3,7 +3,7 @@ import {
   type ProzesskostenhilfePDF,
 } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
-import { fillUnterhaltsanspruch } from "~/domains/prozesskostenhilfe/pdf/C_unterhaltspflichtige_person";
+import { fillUnterhaltsanspruch } from "../C_unterhaltspflichtige_person";
 
 let pdfParams: ProzesskostenhilfePDF;
 const userData: ProzesskostenhilfeFormularContext = {
