+++ app/domains/prozesskostenhilfe/services/pdf/__test__/util.test.ts
@@ -1,4 +1,4 @@
-import { getTotalMonthlyFinancialEntries } from "~/domains/prozesskostenhilfe/pdf/util";
+import { getTotalMonthlyFinancialEntries } from "../util";
 
 describe("utility functions", () => {
   describe("getTotalMonthlyFinancialEntries", () => {
