+++ app/domains/prozesskostenhilfe/services/pdf/index.ts
@@ -1,24 +1,26 @@
 import type { ProzesskostenhilfePDF } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import { getProzesskostenhilfeParameters } from "data/pdf/prozesskostenhilfe/prozesskostenhilfe.generated";
 import type { ProzesskostenhilfeFormularContext } from "~/domains/prozesskostenhilfe/formular";
-import { fillRechtsschutzversicherung } from "~/domains/prozesskostenhilfe/pdf/B_rechtsschutzversicherung";
-import { fillUnterhaltAngehoerige } from "~/domains/prozesskostenhilfe/pdf/D_angehoerige";
-import { fillOwnBruttoEinnahmen } from "~/domains/prozesskostenhilfe/pdf/E_bruttoEinnahmen/bruttoEinnahmen_eigenes";
-import { fillBruttoEinnahmenPartner } from "~/domains/prozesskostenhilfe/pdf/E_bruttoEinnahmen/bruttoEinnahmen_partner";
-import { fillEigentum } from "~/domains/prozesskostenhilfe/pdf/G_eigentum";
-import { fillGrundvoraussetzungen } from "~/domains/prozesskostenhilfe/pdf/grundvoraussetzungen";
 import { fillPerson } from "./A_person";
 import { fillUnterhaltsanspruch } from "./C_unterhaltspflichtige_person";
 import { fillAbzuege } from "./F_abzuege";
 import { fillWohnkosten } from "./H_wohnkosten";
 import { fillZahlungsverpflichtungen } from "./I_zahlungsverpflichtungen";
 import { fillBelastungen } from "./J_belastungen";
-import { appendAttachment } from "../../../services/pdf/appendAttachment";
-import FormAttachment from "../../../services/pdf/attachment/FormAttachment";
-import { pdfFromReact } from "../../../services/pdf/attachment/pdfFromReact";
-import type { PdfFillFunction } from "../../../services/pdf/fillOutFunction";
-import { pdfFillReducer } from "../../../services/pdf/fillOutFunction";
-import { fillPdf } from "../../../services/pdf/fillPdf.server";
+import {
+  PdfFillFunction,
+  pdfFillReducer,
+} from "~/services/pdf/fillOutFunction";
+import { fillRechtsschutzversicherung } from "./B_rechtsschutzversicherung";
+import { fillGrundvoraussetzungen } from "./grundvoraussetzungen";
+import { fillUnterhaltAngehoerige } from "./D_angehoerige";
+import { fillOwnBruttoEinnahmen } from "./E_bruttoEinnahmen/bruttoEinnahmen_eigenes";
+import { fillBruttoEinnahmenPartner } from "./E_bruttoEinnahmen/bruttoEinnahmen_partner";
+import { fillEigentum } from "./G_eigentum";
+import { fillPdf } from "~/services/pdf/fillPdf.server";
+import { appendAttachment } from "~/services/pdf/appendAttachment";
+import { pdfFromReact } from "~/services/pdf/attachment/pdfFromReact";
+import FormAttachment from "~/services/pdf/attachment/FormAttachment";
 export { getProzesskostenhilfeParameters };
 
 export type PkhPdfFillFunction = PdfFillFunction<
